refactor(node): introduce Has* accessor traits and incentive API
Replace monolithic HoprNodeChainOperations and HoprNodeTicketOperations
with composable Has* accessor traits (HasChainApi, HasNetworkView,
HasGraphView, HasTransportApi, HasTicketManagement). Each accessor
exposes a typed component reference and a sync status() method.

Composed traits are blanket-implemented over accessor combinations:
- HoprIncentiveOperations: channels + balances + ticket redemption
- HoprNodeNetworkOperations: pure NetworkView delegation
- HoprChainKeyOperationsExt: chain key resolution

Also adds:
- ComponentStatus enum for per-component health reporting
- TransportOperations trait (minimal transport surface)
- HoprGraphReadApi (read-only graph composite)
